Add JSON storage implementation for tasks and categories
- Introduced a new `JsonStorage` struct to handle persistent storage using JSON files.
- Implemented `save` and `load` methods for storing and retrieving `StorageData`, ensuring data integrity through verification checks.
- Added unit tests to validate the functionality of JSON storage, including tests for saving, loading, and data integrity.
- Updated the `storage` module to include the new JSON storage functionality and adjusted the test suite accordingly.
